✨: ModalProvider를 통한  FrequentlyVisited에 ContextAPI 세팅
