+++ src/components/common/Modal/ModalBase/ModalBase.tsx
@@ -1,49 +1,29 @@
 import React from "react";
 
+export default function ModalBase({
+  isOpen,
+  closeModal,
+  title,
+  children,
+}: ModalBaseProps) {
+  if (!isOpen) return null;
+
+  return (
+    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-50">
+      <div className="relative bg-white w-[90%] max-w-md rounded-lg shadow-lg">
+        <div className="flex justify-between items-center px-4 py-2 border-b">
+          <h3 className="text-lg font-bold">{title}</h3>
+          <button onClick={closeModal}>X</button>
+        </div>
+        <div className="p-4">{children}</div>
+      </div>
+    </div>
+  );
+}
+
 interface ModalBaseProps {
   isOpen: boolean;
   closeModal: () => void;
   title: string;
   children: React.ReactNode;
 }
-
-const ModalBase: React.FC<ModalBaseProps> = ({ isOpen, closeModal, title, children }) => {
-  if (!isOpen) return null;
-
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center bg-gray-800 bg-opacity-50 backdrop-blur-sm transition-opacity duration-300"
-      role="dialog"
-      aria-modal="true"
-    >
-      {/* 모달 컨테이너 */}
-      <div className="relative w-full max-w-md bg-white rounded-2xl shadow-xl">
-        {/* 모달 헤더 */}
-        <div className="px-6 py-4 bg-gray-100 rounded-t-2xl flex justify-between items-center">
-          <h3 className="text-lg font-bold text-gray-900">{title}</h3>
-          <button
-            type="button"
-            className="text-gray-400 hover:text-gray-600 focus:outline-none"
-            onClick={closeModal}
-          >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              fill="none"
-              viewBox="0 0 24 24"
-              strokeWidth="2"
-              stroke="currentColor"
-              className="h-6 w-6"
-            >
-              <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-
-        {/* 모달 본문 */}
-        <div className="p-6">{children}</div>
-      </div>
-    </div>
-  );
-};
-
-export default ModalBase;
