+++ src/hooks/modal/useModal.ts
@@ -1,9 +1,19 @@
-import { useState } from "react";
+import { useContext } from "react";
+import { ModalContext } from "../../components/common/Modal/ModalBase/ModalContext";
 
 export default function useModal() {
-  const [isOpen, setIsOpen] = useState(false);
-  const openModal = () => setIsOpen(true);
-  const closeModal = () => setIsOpen(false);
+  const context = useContext(ModalContext);
 
-  return { isOpen, openModal, closeModal };
+  if (!context) {
+    throw new Error("useModal must be used within a ModalProvider");
+  }
+
+  return {
+    isModalBaseOpen: context.isModalBaseOpen,
+    isModalWithPaginationOpen: context.isModalWithPaginationOpen,
+    openModalBase: context.openModalBase,
+    closeModalBase: context.closeModalBase,
+    openModalWithPagination: context.openModalWithPagination,
+    closeModalWithPagination: context.closeModalWithPagination,
+  };
 }
